Use slower clock for live-io demo
Adding the Pcam 5C video pipeline seems to be make timing harder
to meet. To account for this, the default clock on the HDMI input
demo is turned down to 100 MHz. It should still be possible to
successfully build at higher speeds, but it will be up to the
user to mess with these settings and see how fast they can get
it.